requirements.txt
flask==2.2.5
flask-sqlalchemy==3.0.2
python-dotenv==0.21.1
psycopg2-binary==2.9.5
gunicorn==20.1.0
werkzeug==2.2.3 